Move types near features
